+++ photutils/detection/core.py
@@ -4,9 +4,11 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 import numpy as np
-from astropy.table import Table
+from astropy.table import Column, Table
 from astropy.convolution import Kernel2D
 from astropy.stats import sigma_clipped_stats
+from ..morphology import cutout_footprint, fit_2dgaussian
+from ..utils.wcs_helpers import pixel_to_icrs_coords
 
 
 __all__ = ['detect_threshold', 'detect_sources', 'find_peaks']
@@ -232,7 +234,7 @@ def detect_sources(data, threshold, npixels, filter_kernel=None,
     elif connectivity == 8:    # e.g., SExtractor
         selem = ndimage.generate_binary_structure(2, 2)
     else:
-        raise ValueError('Invalid selem={0}.  '
+        raise ValueError('Invalid connectivity={0}.  '
                          'Options are 4 or 8'.format(connectivity))
 
     objlabels, nobj = ndimage.label(image, structure=selem)
@@ -250,19 +252,31 @@ def detect_sources(data, threshold, npixels, filter_kernel=None,
     return objlabels
 
 
-def find_peaks(data, threshold, min_separation=2, exclude_border=True,
-               segment_image=None, npeaks=np.inf, footprint=None):
+def find_peaks(data, threshold, box_size=3, footprint=None, mask=None,
+               border_width=None, npeaks=np.inf, subpixel=False, error=None,
+               wcs=None):
     """
     Find local peaks in an image that are above above a specified
     threshold value.
 
-    Peaks are the local maxima above the ``threshold`` in a region of
-    ``(2 * min_separation) + 1`` (i.e., peaks are separated by at least
-    ``min_separation`` pixels).
+    Peaks are the maxima above the ``threshold`` within a local region.
+    The regions are defined by either the ``box_size`` or ``footprint``
+    parameters.  ``box_size`` defines the local region around each pixel
+    as a square box.  ``footprint`` is a boolean array where `True`
+    values specify the region shape.
 
-    If peaks are flat (i.e., multiple adjacent pixels have identical
-    intensities), then the coordinates of all such pixels are returned,
-    even if they are not separated by at least ``min_separation``.
+    If multiple pixels within a local region have identical intensities,
+    then the coordinates of all such pixels are returned.  Otherwise,
+    there will be only one peak pixel per local region.  Thus, the
+    defined region effectively imposes a minimum separation between
+    peaks (unless there are identical peaks within the region).
+
+    When using subpixel precision (``subpixel=True``), then a cutout of
+    the specified ``box_size`` or ``footprint`` will be taken centered
+    on each peak and fit with a 2D Gaussian (plus a constant).  In this
+    case, the fitted local centroid and peak value (the Gaussian
+    amplitude plus the background constant) will also be returned in the
+    output table.
 
     Parameters
     ----------
@@ -272,22 +286,27 @@ def find_peaks(data, threshold, min_separation=2, exclude_border=True,
     threshold : float
         The data value to be used for the detection threshold.
 
-    min_separation : int, optional
-        Minimum number of pixels separating peaks (i.e., peaks are
-        separated by at least ``min_separation`` pixels).  To find the
-        maximum number of peaks, use ``min_separation=1``.  If
-        ``min_separation`` is not an integer, then it will be truncated.
+    box_size : scalar or tuple, optional
+        The size of the local region to search for peaks at every point
+        in ``data``.  If ``box_size`` is a scalar, then the region shape
+        will be ``(box_size, box_size)``.  Either ``box_size`` or
+        ``footprint`` must be defined.  If they are both defined, then
+        ``footprint`` overrides ``box_size``.
 
-    exclude_border : bool, optional
-        If `True`, exclude peaks within ``min_separation`` from the
-        border of the image as well as from each other.
+    footprint : `~numpy.ndarray` of bools, optional
+        A boolean array where `True` values describe the local footprint
+        region within which to search for peaks at every point in
+        ``data``.  ``box_size=(n, m)`` is equivalent to
+        ``footprint=np.ones((n, m))``.  Either ``box_size`` or
+        ``footprint`` must be defined.  If they are both defined, then
+        ``footprint`` overrides ``box_size``.
 
-    segment_image : `~numpy.ndarray` (int), optional
-        If provided, then search for peaks located only within the
-        labeled regions of a 2D segmentation image, where sources are
-        marked by different positive integer values.  In the
-        segmentation image a value of zero is reserved for the
-        background.  ``segment_image`` must have the same shape as
+    mask : array_like, bool, optional
+        A boolean mask with the same shape as ``data``, where a `True`
+        value indicates the corresponding element of ``data`` is masked.
+
+    border_width : bool, optional
+        The width in pixels to exclude around the border of the
         ``data``.
 
     npeaks : int, optional
@@ -295,39 +314,111 @@ def find_peaks(data, threshold, min_separation=2, exclude_border=True,
         detected peaks exceeds ``npeaks``, the peaks with the highest
         peak intensities will be returned.
 
-    footprint : `~numpy.ndarray` of bools, optional
-        A boolean array where `True` values describe the local footprint
-        region within which to search for peaks at every point in
-        ``data``.  Overrides ``min_separation``, except for border
-        exclusion when ``exclude_border=True``.
+    subpixel : bool, optional
+        If `True`, then a cutout of the specified ``box_size`` or
+        ``footprint`` will be taken centered on each peak and fit with a
+        2D Gaussian (plus a constant).  In this case, the fitted local
+        centroid and peak value (the Gaussian amplitude plus the
+        background constant) will also be returned in the output table.
+
+    error : array_like, optional
+        The 2D array of the 1-sigma errors of the input ``data``.
+        ``error`` is used only to weight the 2D Gaussian fit performed
+        when ``subpixel=True``.
+
+    wcs : `~astropy.wcs.WCS`
+        The WCS transformation to use to convert from pixel coordinates
+        to ICRS world coordinates.  If `None`, then the world
+        coordinates will not be returned in the output
+        `~astropy.table.Table`.
 
     Returns
     -------
     output : `~astropy.table.Table`
         A table containing the x and y pixel location of the peaks and
-        their values.
+        their values.  If ``subpixel=True``, then the table will also
+        contain the local centroid and fitted peak value.
     """
 
-    if segment_image is not None:
-        if segment_image.shape != data.shape:
-            raise ValueError('segment_image and data must have the same '
-                             'shape')
+    from scipy import ndimage
 
-    from skimage.feature import peak_local_max
-    coords = peak_local_max(data, min_distance=int(min_separation),
-                            threshold_abs=threshold, threshold_rel=0.0,
-                            exclude_border=exclude_border, indices=True,
-                            num_peaks=npeaks, footprint=footprint,
-                            labels=segment_image)
-    y_peaks, x_peaks = coords[:, 0], coords[:, 1]
+    if np.all(data == data.flat[0]):
+        return []
+
+    if footprint is not None:
+        data_max = ndimage.maximum_filter(data, footprint=footprint,
+                                          mode='constant', cval=0.0)
+    else:
+        data_max = ndimage.maximum_filter(data, size=box_size,
+                                          mode='constant', cval=0.0)
+
+    peak_goodmask = (data == data_max)    # good pixels are True
+
+    if mask is not None:
+        mask = np.asanyarray(mask)
+        if data.shape != mask.shape:
+            raise ValueError('data and mask must have the same shape')
+        peak_goodmask = np.logical_and(peak_goodmask, ~mask)
+
+    if border_width is not None:
+        for i in range(peak_goodmask.ndim):
+            peak_goodmask = peak_goodmask.swapaxes(0, i)
+            peak_goodmask[:border_width] = False
+            peak_goodmask[-border_width:] = False
+            peak_goodmask = peak_goodmask.swapaxes(0, i)
+
+    peak_goodmask = np.logical_and(peak_goodmask, (data > threshold))
+    y_peaks, x_peaks = peak_goodmask.nonzero()
     peak_values = data[y_peaks, x_peaks]
 
-    if coords.shape[0] > npeaks:
-        # NOTE: num_peaks is ignored by peak_local_max() if labels are input
+    if len(x_peaks) > npeaks:
         idx = np.argsort(peak_values)[::-1][:npeaks]
         x_peaks = x_peaks[idx]
         y_peaks = y_peaks[idx]
         peak_values = peak_values[idx]
 
-    return Table((x_peaks, y_peaks, peak_values),
-                 names=('x_peak', 'y_peak', 'peak_value'))
+    if subpixel:
+        x_centroid, y_centroid = [], []
+        fit_peak_values = []
+        for (y_peak, x_peak) in zip(y_peaks, x_peaks):
+            rdata, rmask, rerror, slc = cutout_footprint(
+                data, (x_peak, y_peak), box_size=box_size,
+                footprint=footprint, mask=mask, error=error)
+            gaussian_fit = fit_2dgaussian(rdata, mask=rmask, error=rerror)
+            if gaussian_fit is None:
+                x_cen, y_cen, fit_peak_value = np.nan, np.nan, np.nan
+            else:
+                x_cen = slc[1].start + gaussian_fit.x_mean_1.value
+                y_cen = slc[0].start + gaussian_fit.y_mean_1.value
+                fit_peak_value = (gaussian_fit.amplitude_0.value +
+                                  gaussian_fit.amplitude_1.value)
+            x_centroid.append(x_cen)
+            y_centroid.append(y_cen)
+            fit_peak_values.append(fit_peak_value)
+
+        columns = (x_peaks, y_peaks, peak_values, x_centroid, y_centroid,
+                   fit_peak_values)
+        names = ('x_peak', 'y_peak', 'peak_value', 'x_centroid', 'y_centroid',
+                 'fit_peak_value')
+    else:
+        columns = (x_peaks, y_peaks, peak_values)
+        names = ('x_peak', 'y_peak', 'peak_value')
+
+    table = Table(columns, names=names)
+
+    if wcs is not None:
+        icrs_ra_peak, icrs_dec_peak = pixel_to_icrs_coords(x_peaks, y_peaks,
+                                                           wcs)
+        table.add_column(Column(icrs_ra_peak, name='icrs_ra_peak'), index=2)
+        table.add_column(Column(icrs_dec_peak, name='icrs_dec_peak'), index=3)
+
+        if subpixel:
+            icrs_ra_centroid, icrs_dec_centroid = pixel_to_icrs_coords(
+                x_centroid, y_centroid, wcs)
+            idx = table.colnames.index('y_centroid')
+            table.add_column(Column(icrs_ra_centroid,
+                                    name='icrs_ra_centroid'), index=idx+1)
+            table.add_column(Column(icrs_dec_centroid,
+                                    name='icrs_dec_centroid'), index=idx+2)
+
+    return table